+++ app/controllers/profiles_controller.rb
@@ -2,6 +2,7 @@
 
 class ProfilesController < ApplicationController
   before_action :find_end_time, only: %i[show]
+  before_action :find_vacations, only: %i[show]
 
   def show; end
 
@@ -19,5 +20,13 @@ class ProfilesController < ApplicationController
       .subscriptions
       .where('start_time < :date AND end_time > :date', date: DateTime.now.new_offset(0))
       .maximum(:end_time)
+  end
+
+  def find_vacations
+    @vacations =
+      current_user
+      .vacations
+      .order(start_time: :desc)
+      .load_async
   end
 end